+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/logging/LogMessageKeys.java
@@ -152,6 +152,8 @@ public enum LogMessageKeys {
     CONTINUED_BUILD("continued_build"),
     INDEXING_METHOD("indexing_method"),
     ALLOW_REPAIR("allow_repair"),
+    INDEXING_POLICY("indexing_policy"),
+    INDEXING_POLICY_DESIRED_ACTION("indexing_policy_desired_action"),
 
     // comparisons
     COMPARISON_VALUE("comparison_value"),